app/path_curves.[ch] app/tools/gimpdrawtool.c app/tools/gimppathtool.[ch]
2001-04-11  Simon Budig  <simon@gimp.org>

        * app/path_curves.[ch]
        * app/tools/gimpdrawtool.c
        * app/tools/gimppathtool.[ch]
        * app/tools/path_tool.[ch]
        * app/tools/path_toolP.h

        At least now it looks as if it could do something sometimes...

+++ app/tools/gimppathtool.h
@@ -38,8 +38,8 @@ struct _GimpPathTool
   GimpDrawTool     parent_instance;
 
   gint         click_type;      /* where did the user click?         */
-  gint         click_x;         /* X-coordinate of the click         */
-  gint         click_y;         /* Y-coordinate of the click         */
+  gdouble      click_x;         /* X-coordinate of the click         */
+  gdouble      click_y;         /* Y-coordinate of the click         */
   gint         click_halfwidth;
   guint        click_modifier;  /* what modifiers were pressed?      */
   NPath       *click_path;      /* On which Path/Curve/Segment       */
